% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{all_questions}
\alias{all_questions}
\alias{all_solutions}
\title{Format a problem/solution into markdown}
\usage{
all_questions(problem, listch = "1.", prepend_question = TRUE, spacer = "\\n")

all_solutions(problem, listch = "1", prepend_solution = TRUE)
}
\arguments{
\item{problem}{a fully compiled problem.}

\item{listch}{string to prepend subquestions. Defaults to \code{"1."} (numbered list).}

\item{prepend_question, prepend_solution}{logical (default \code{TRUE}) whether to include the \code{question}/\code{solution} field before listing the \code{subquestions}/\code{subsolutions}.}

\item{spacer}{string (default \code{"\\n"}) to append to subquestions.}
}
\value{
string containing the compiled problem/solution formatted into markdown.
}
\description{
Format a problem/solution into markdown
}
\examples{
all_questions(simple_prop(seed=1))
all_solutions(simple_prop(seed=1))

all_questions(simple_prop_percent(seed=1))
all_solutions(simple_prop_percent(seed=1))
}
