% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{build_problemset}
\alias{build_problemset}
\title{Container for multiple problem templates}
\usage{
build_problemset(template_list, prob = NULL, title = "", label = title)
}
\arguments{
\item{template_list}{list of problem templates.}

\item{prob}{optional vector of weights for use when sampling problems from the problem set. Defaults to \code{NULL}, representing a uniform distribution.}

\item{title}{optional string used as a short title description of the whole set.}

\item{label}{optional shorts label for the problem set. Defaults to the same value as \code{title}.}
}
\value{
a problemset object.
}
\description{
Container for multiple problem templates
}
\examples{
percent_problems = build_problemset(list(
  simple_percent,
  simple_percent_successes
), title="Percents")
}
