% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_propagation.R
\name{ceilingk}
\alias{ceilingk}
\alias{floork}
\title{Ceiling and floor with different precision}
\usage{
ceilingk(x, k = 0)

floork(x, k = 0)
}
\arguments{
\item{x}{a numeric vector of values to be rounded up or down.}

\item{k}{integer indicating the number of decimal places to be rounded to. Default of 0}
}
\value{
a numeric vector of rounded values.
}
\description{
\code{ceilingk} rounds a numeric vector \code{x} up to \code{k} decimal places.

\code{floork} rounds a numeric vector \code{x} down to \code{k} decimal places.
}
\examples{
ceilingk(3.14159, 2)
floork(3.14159, 2)
}
