% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{comp}
\alias{comp}
\title{Compare values and summarize as (in)equality symbols}
\usage{
comp(a, b)
}
\arguments{
\item{a, b}{values to compare.}
}
\value{
character "=" if equal, ">" if \code{a} is greater than \code{b}, "<" if \code{a} is less than \code{b}.
}
\description{
Compare values and summarize as (in)equality symbols
}
\examples{
comp(1,1)
comp(1,2)
comp(2,1)
}
