% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{compile_figure}
\alias{compile_figure}
\title{Pre-specify plotting function arguments}
\usage{
compile_figure(func, vars)
}
\arguments{
\item{func}{function Plotting function with single argument taking a list of
plotting variables.}

\item{vars}{listish List or environment containing variables to make
available for string interpolation. Defaults to \code{parent.frame()}.}
}
\value{
A function with no arguments that uses the original function on the
variables in vars.
}
\description{
Pre-specify plotting function arguments
}
\examples{
vars <- list(x = 1:5, y = 5:1)
func <- function(args) plot(args$x, args$y, pch=19)
func_noargs <- compile_figure(func, vars)
func_noargs()
}
