% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{compile_text}
\alias{compile_text}
\title{LaTeX-friendly string interpolation}
\usage{
compile_text(
  text,
  vars = parent.frame(),
  open = "<<",
  close = ">>",
  trim = FALSE
)
}
\arguments{
\item{text}{string Input template string.}

\item{vars}{listish List or environment containing variables to make available for string interpolation. Defaults to \code{parent.frame()}.}

\item{open}{string Opening delimiter. Defaults to "<<".}

\item{close}{string Closing delimiter. Defaults to ">>".}

\item{trim}{logical Whether to trim \code{text} with \code{\link[glue:trim]{glue::trim()}} or not. Unlike \code{\link[glue:glue]{glue::glue()}}, defaults to \code{FALSE}.}
}
\value{
A glue object.
}
\description{
Wrapper for the \code{\link[glue:glue]{glue::glue()}} function with more LaTeX-friendly and
raw-string-friendly defaults.
}
\examples{
details <- list(name = "Michael", subject="statistics")
compile_text("My name is <<name>>, and I teach <<subject>>.", vars = details)

setup <- list(ex = 3)
compile_text(
  r"(The derivative of $\pm C x^<<ex>>$ is $\pm <<ex>> C x^<<ex-1>>$)",
  vars = setup)
}
