% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{dotplot}
\alias{dotplot}
\title{Dot plot (histograms)}
\usage{
dotplot(
  x,
  breaks = "Sturges",
  xlim = NULL,
  ylim = NULL,
  plot = TRUE,
  xlab = NULL,
  ylab = "Count",
  ...
)
}
\arguments{
\item{x}{vector of values to be plotted.}

\item{breaks}{passed through to \code{\link[=hist]{hist()}}.}

\item{xlim, ylim}{the range of x and y values to visualize.}

\item{plot}{logical if \code{TRUE} (default) the dot plot is plotted.}

\item{xlab, ylab}{axis labels.}

\item{...}{further arguments passed through to \code{\link[=plot]{plot()}}.}
}
\value{
histogram object underlying the dot plot.
}
\description{
Dot plot (histograms)
}
\examples{
dotplot(c(1,2,3,3,5,5,5,5,6,7,7,9))
}
