% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{format_problem}
\alias{format_problem}
\alias{format_solution}
\title{Format a problem/solution into markdown with figures and tables}
\usage{
format_problem(prob, pts = NULL, spacer = "")

format_solution(prob)
}
\arguments{
\item{prob}{a fully compiled problem.}

\item{pts}{optional number of points.}

\item{spacer}{string to be passed through to \code{all_questions()}/\code{all_solutions()}.}
}
\description{
Format a problem/solution into markdown with figures and tables
}
\examples{
format_problem(simple_prop(seed=1))
format_solution(simple_prop(seed=1))
}
