% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_problem.R
\name{format_problems}
\alias{format_problems}
\title{Format a list of problems and (optionally) solutions}
\usage{
format_problems(probs, pts_list = NULL, show_sol = FALSE)
}
\arguments{
\item{probs}{list of compiled problems.}

\item{pts_list}{optional vector of points.}

\item{show_sol}{logical (default \code{FALSE}) whether to show solutions.}
}
\description{
Format a list of problems and (optionally) solutions
}
\examples{
format_problems(list(simple_prop(seed=1), simple_percent(seed=2)))
}
