% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{greater_less}
\alias{greater_less}
\title{Expand inequality symbol to word}
\usage{
greater_less(gtlt, capitalize = FALSE)
}
\arguments{
\item{gtlt}{">" or "<"}

\item{capitalize}{logical whether to capitalize the first letter.}
}
\value{
string containing "greater" or "less", potentially capitalized.
}
\description{
Expand inequality symbol to word
}
\examples{
greater_less("<")
}
