% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{indep_tab}
\alias{indep_tab}
\title{Generate independent two-way tables}
\usage{
indep_tab(n, rprobs, cprobs, dnames = NULL, totals = TRUE)
}
\arguments{
\item{n}{integer sample size.}

\item{rprobs, cprobs}{vectors of row- and column-probabilities.}

\item{dnames}{list of dimension names. See \code{\link[=dimnames]{dimnames()}}.}

\item{totals}{logical (default \code{TRUE}) whether to include row and column totals.}
}
\value{
matrix containing a two way table.
}
\description{
Generate independent two-way tables
}
\examples{
indep_tab(100, c(0.5,0.5), c(0.3,0.4,0.3),
  dnames=list(Sex=c('M','F'), class=c('A','B','C')), totals=TRUE)
}
