% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{is_neg}
\alias{is_neg}
\title{Check for negative values}
\usage{
is_neg(x)
}
\arguments{
\item{x}{vector of values to check for negativeness. Run through \link{as.numeric} first.}
}
\value{
logcical vector of \code{TRUE}s for negative values and \code{FALSE}s for nonnegative values.
}
\description{
Check for negative values
}
\examples{
is_neg(-1)

is_neg(c(10, -10))
}
