% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{lgd}
\alias{lgd}
\title{Expand inequality symbol to word}
\usage{
lgd(ineq)
}
\arguments{
\item{ineq}{string ">" or "<"}
}
\value{
string "less" if \code{ineq} is "<", "greater" if \code{ineq} is ">", "different" if \code{ineq} is anything else.
}
\description{
Expand inequality symbol to word
}
\examples{
lgd("<")
lgd(">")
lgd(neq())
}
