% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{lower_tail}
\alias{lower_tail}
\title{Determine whether a p-value calculation should use lower.tail.}
\usage{
lower_tail(ineq, z)
}
\arguments{
\item{ineq}{string "<" gives a "less than" test, ">" gives a "greater than" test, anything else gives a "two-sided" test.}

\item{z}{numeric test statistic.}
}
\value{
logical whether p-value calls to \code{\link[stats:Normal]{stats::pnorm()}} or \code{\link[stats:TDist]{stats::pt()}} should use \code{lower.tail}.
}
\description{
Determine whether a p-value calculation should use lower.tail.
}
\examples{
lower_tail("<", 1.5)
lower_tail(neq(), -1.5)
}
