% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{neq}
\alias{neq}
\title{Shortcut for LaTeX not equal character}
\usage{
neq()
}
\value{
string \code{r"(\\neq)"}
}
\description{
Shortcut for LaTeX not equal character
}
\examples{
compile_text(r"($x <<neq()>> 0$)")
}
