% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{points_block}
\alias{points_block}
\title{Generate table of quiz points}
\usage{
points_block(pts)
}
\arguments{
\item{pts}{vector of point values.}
}
\value{
string containing a points block as rendered by \code{\link[knitr:kable]{knitr::kable()}}.
}
\description{
Generate table of quiz points
}
\examples{
points_block(c(10,15,20))
}
