% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rdist}
\alias{rdist}
\title{Random finite probability distribution}
\usage{
rdist(k, concentration = 1)
}
\arguments{
\item{k}{integer; the size of the probability distribution.}

\item{concentration}{numeric; acts as the concentration parameter to the symmetric Dirichlet distribution.}
}
\value{
numeric vector of probabilities.
}
\description{
Generates a finite-dimensional probability distribution according to a symmetric \href{https://en.wikipedia.org/wiki/Dirichlet_distribution#Special_cases}{Dirichlet distribution}.
}
\examples{
rdist(5, 1)
}
