% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repractice-package.R
\docType{package}
\name{repractice-package}
\alias{repractice}
\alias{repractice-package}
\title{repractice: A Framework for Reproducible Practice Problems}
\description{
A portable framework for randomizable logic and markdown templating, with a focus on generating combinatorially variable practice problems and solutions.
}
\keyword{internal}
