% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_interface.R
\name{repractice_shiny_tabbed}
\alias{repractice_shiny_tabbed}
\title{Tabbed Shiny interface for multiple problem sets}
\usage{
repractice_shiny_tabbed(problemset_list)
}
\arguments{
\item{problemset_list}{list of problemsets.}
}
\value{
A \code{\link[=shinyApp]{shinyApp()}} object.
}
\description{
Tabbed Shiny interface for multiple problem sets
}
\examples{
prop_problems = build_problemset(list(
  simple_prop,
  pet_prop,
  simple_successes,
  dotplot_visual_prop
), title="Proportions")

percent_problems = build_problemset(list(
  simple_percent,
  simple_percent_successes
), title="Percents")

\dontrun{
repractice_shiny_tabbed(list(prop_problems, percent_problems))
}
}
