% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_problems.R
\name{simple_prop}
\alias{simple_prop}
\alias{pet_prop}
\alias{simple_percent}
\alias{simple_prop_percent}
\alias{simple_successes}
\alias{simple_percent_successes}
\alias{dotplot_visual_prop}
\title{Simple proportion/percent problem templates.}
\usage{
simple_prop(seed = NULL)

pet_prop(seed = NULL)

simple_percent(seed = NULL)

simple_prop_percent(seed = NULL)

simple_successes(seed = NULL)

simple_percent_successes(seed = NULL)

dotplot_visual_prop(seed = NULL)
}
\arguments{
\item{seed}{(optional) integer; pseudorandom seed used with \code{\link[withr:with_seed]{withr::local_seed()}} inside the problem.}
}
\value{
compiled problem
}
\description{
Collection of simple example problems involving proportions and/or percents.
}
\examples{
simple_prop(seed=1)
pet_prop(seed=2)
simple_percent(seed=3)
simple_prop_percent(seed=4)
simple_successes(seed=5)
simple_percent_successes(seed=6)
dotplot_visual_prop(seed=7)
}
