% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{t_ineq_pval}
\alias{t_ineq_pval}
\title{Calculate p-value for a t test.}
\usage{
t_ineq_pval(t, df, ineq)
}
\arguments{
\item{t}{numeric t test statistic.}

\item{df}{integer degrees of freedom.}

\item{ineq}{string "<" gives a "less than" test, ">" gives a "greater than" test, anything else gives a "two-sided" test.}
}
\value{
numeric p-value for the t test.
}
\description{
Calculate p-value for a t test.
}
\examples{
t_ineq_pval(0.5, 10, ">")
t_ineq_pval(2.0, 30, neq())
}
