% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{vspace}
\alias{vspace}
\title{Shortcut for LaTeX vertical spacing}
\usage{
vspace(v)
}
\arguments{
\item{v}{string to insert into \verb{\\vspace\{...\}}.}
}
\value{
string for a LaTeX vertical space \verb{\\vspace\{...\}}
}
\description{
Shortcut for LaTeX vertical spacing
}
\examples{
vspace("30pt")
}
