% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_helpers.R
\name{wrapif}
\alias{wrapif}
\title{Wrap in delimiters under certain conditions}
\usage{
wrapif(x, cond = is_neg, left = "(", right = ")")
}
\arguments{
\item{x}{object under consideration.}

\item{cond}{function returning a logical value.}

\item{left}{string left delimiter.}

\item{right}{string right delimiter.}
}
\value{
string containing the potentially wrapped input.
}
\description{
Conditional wrapping function. Main use case is enclosing negative values in parentheses.
}
\examples{
wrapif(5)
wrapif(-5)
}
