% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{z_ineq_pval}
\alias{z_ineq_pval}
\title{Calculate p-value for a z-test.}
\usage{
z_ineq_pval(z, ineq)
}
\arguments{
\item{z}{numeric z-test statistic.}

\item{ineq}{string "<" gives a "less than" test, ">" gives a "greater than" test, anything else gives a "two-sided" test.}
}
\value{
numeric p-value for the z-test.
}
\description{
Calculate p-value for a z-test.
}
\examples{
z_ineq_pval(0.5, ">")
z_ineq_pval(2.0, neq())
}
